backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pillow
flask-cors
